A Major update

+++ src/main/java/com/projectForum/Services/DeleteService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.projectForum.ControlPanel.SearchUserForm;
 import com.projectForum.forum.Forum;
 import com.projectForum.forum.ForumRepository;
 import com.projectForum.post.Post;
@@ -13,6 +14,13 @@ import com.projectForum.topic.Topic;
 import com.projectForum.topic.TopicRepository;
 import com.projectForum.user.User;
 import com.projectForum.user.UserRepository;
+
+
+//TODO TEST THIS FILE
+/* ###########################################################
+* 		WARNING: THIS SERVICE FILE ISN'T TESTED YET!
+* ###########################################################*/
+
 
 /**
  *  This Class will use as a service to all deletion option in the application */
@@ -34,12 +42,83 @@ public class DeleteService {
 		this.forumRepo = forumRepo;
 	}
 	
-	/** This method will be used to delete a user*/
-	public void deleteUser(int userId) {
-		// TODO find a solution to delete user but not topics and posts
-		User user = userRepo.findUserById(userId);
+	/** This method will delete an exists user.
+	 * 	However it won't delete it posts and topics, but will attached user's posts and topics
+	 * 	to an Dummy User.
+	 * 
+	 * @param User
+	 * */
+	public void deleteUserKeepActivity(User user) {
+		
+		User dummyUser = userRepo.findByUsername("Unknown");
+		
+		// Making sure that dummy user won't be deleted or Admin user
+		if(dummyUser.equals(user) || user.getRoles().iterator().next().getName().equals("ADMIN"))
+			return;
+		
+		// Removing Role (many To many, need to be removed)
+		user.removeRole();
+		
+		// Change the topic and post to dummy owner.
+		List<Post>	userPosts	=	postRepo.findPostsByUser(user);
+		List<Topic>	userTopics	=	topicRepo.findTopicsByUser(user);
+		
+		// posts:
+		if(!userPosts.isEmpty())
+			for (int i=0; i<userPosts.size(); i++) {
+				userPosts.get(i).setUser(dummyUser);
+			}
+		// topics:
+		if(!userTopics.isEmpty())
+			for (int i=0; i<userTopics.size(); i++) {
+				userTopics.get(i).setUser(dummyUser);
+			}
 		
 		userRepo.delete(user);
+	}
+	
+	/** This method will delete an exists user.
+	 * 	it will also delete it posts and topics.
+	 * 	to an Dummy User.
+	 * 
+	 * @param User
+	 * */
+	public void deleteUserDontKeepActivity(User user) {
+		List<Post>	userPosts	=	postRepo.findPostsByUser(user);
+		List<Topic>	userTopics	=	topicRepo.findTopicsByUser(user);
+		
+		// Removing all posts and topics
+		this.deletePosts(userPosts);
+		this.deleteTopics(userTopics);
+		// Removing user
+		userRepo.delete(user);
+	}
+	/** This method will delete an exists user.
+	 *  It will remove user's data dependent on Keep activity boolean.*/
+	public void deleteUser(SearchUserForm user) {
+			if(user.getKeepActivity())
+				this.deleteUserKeepActivity(user.getUser());
+			else
+				this.deleteUserDontKeepActivity(user.getUser());
+	}
+	/** This method will delete an exists user.
+	 * 	However it won't delete it posts and topics, but will attached user's posts and topics
+	 * 	to an Dummy User.
+	 * 
+	 * @param String
+	 * */
+	public void deleteUser(String username) {
+		this.deleteUserKeepActivity(userRepo.findByUsername(username));
+	}
+	
+	/** This method will delete an exists user.
+	 * 	However it won't delete it posts and topics, but will attached user's posts and topics
+	 * 	to an Dummy User.
+	 * 
+	 * @param int
+	 * */
+	public void deleteUser(int userId) {
+		this.deleteUserKeepActivity(userRepo.findUserById(userId));
 	}
 	
 	/** This method will delete a post by postId
@@ -59,9 +138,11 @@ public class DeleteService {
 	 * @param List<Post>*/
 	
 	public void deletePosts(List<Post> posts) {
-		while(!posts.isEmpty()) {
-			this.deletePost(posts.get(0));
-		}
+
+		if(posts.isEmpty())
+			return;
+		for(int i=0; i<posts.size(); i++)
+			this.deletePost(posts.get(i));
 			
 	}
 	
@@ -90,9 +171,12 @@ public class DeleteService {
 	/** This method will delete list of topics by List<Topic>
 	 *  @param List<Topic>*/
 	public void deleteTopics(List<Topic> topics) {
-		while(!topics.isEmpty()) {
-			this.deleteTopic(topics.get(0));
-		}
+
+		if(topics.isEmpty())
+			return;
+		
+		for (int i=0; i<topics.size(); i++)
+			this.deleteTopic(topics.get(i));
 	}
 	
 	/** This method will delete a forum by forumId
@@ -105,6 +189,9 @@ public class DeleteService {
 	/** This method will delete a forum by forum object
 	 *  @param Forum*/
 	public void deleteForum(Forum forum) {
+		if(forum == null)
+			return;
+		this.updateAllLowerPriority(forum);
 		// Get all topics in Forum
 		List<Topic> topics = topicRepo.findTopicsByForum(forum);
 		
@@ -117,16 +204,38 @@ public class DeleteService {
 	
 	/** This method will delete a list of forums by List<forum>*/
 	public void deleteForums(List<Forum> forums) {
-		while(!forums.isEmpty()) {
-			this.deleteForum(forums.get(0));
-		}
+
+		if(forums.isEmpty())
+			return;
+		for(int i=0; i<forums.size(); i++)
+			this.deleteForum(forums.get(i));
 	}
 	
 	/** This method will delete all the forums + Topics + Posts in database.
-	 * @warning THIS CAN'T BE UNDONE*/
-	public void deleteAll() {
+	 * @warning THIS CAN'T BE UNDONE
+	 * 
+	 * There is a bug in this method, it isn't working.*/
+	public void deleteAllForums() {
+		
 		List<Forum> forums = forumRepo.findAll();
 		this.deleteForums(forums);
 		
+	}
+	
+	/**
+	 * When deleting a forum there is need to update all lower priority */
+	private void updateAllLowerPriority(Forum forum) {
+		final int forumsSize = forumRepo.findByOrderByPriorityAsc().size();
+		
+		// if true then no need to update any other forum's priority
+		if (forum.getPriority() == forumsSize)
+			return;
+		// forum isn't in last priority, need to update all lower priority forums
+		
+		for (int i = forum.getPriority() + 1; i <= forumsSize; i++) {
+			Forum updateForum = forumRepo.findByPriority(i);
+			updateForum.setPriority(i - 1);
+			forumRepo.save(updateForum);
+		}
 	}
 }
